adding delete user function 27

+++ application/routes.py
@@ -22,9 +22,10 @@ def main(user = "No User"):
     
     if deluser.submit.data:
         duser = User.query.filter_by(name=user).first()
-        dpost = Post.query.filter_by(user_id = duser.id).all()
+        if Post.query.filter_by(user = duser).first():
+            dpost = Post.query.filter_by(user = duser).first():
+            db.session.delete(dpost)
         db.session.delete(duser)
-        db.session.delete(dpost)
         db.session.commit()
         return redirect(url_for('login'))
